backend/pyproject.toml
[tool.poetry]
name = "ai-test2"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Pillow = "^9.2.0"
torch= {url = "https://download.pytorch.org/whl/cu116/torch-1.12.1%2Bcu116-cp310-cp310-linux_x86_64.whl"}
transformers = "^4.21.1"
Flask = "^2.2.2"
webargs = "^8.2.0"
#diffusers = "^0.2.2"
diffusers = { git = "https://github.com/huggingface/diffusers.git", rev="e30e1b89d0a85073f8f5338b86445bf8ed892086" }

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
